fix(deployment-location): location is now a variable, no longer uses a data source
using a data source made the deployment want to redeploy the entire cluster if the underlaying resource group had any minor changes to it, making it hard to have a stable deployment with the module. Removing the data source and asking for a location input shouldn't make it harder to use but removes any weird dependency issues.

closes #60

Signed-off-by: Roberth Strand <roberth.strand@crayon.com>