feat(addon_profile): created dynamic block for secrets provider
As the addon block was removed, there are a number of changes to the way this attributes worked. The 'key_vault_secrets_provider' block now doesn't have an attribute called enabled, but by defining the two other attributes it interpretates it as enabled. Made it a dynamic block to avoid it being defined unless these values are set.

Signed-off-by: Roberth Strand <roberth.strand@crayon.com>